# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorBody {
  error: string;
}

export interface ArchiveInput {
  archived: boolean;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  durationDays: number;
  active: boolean;
}

export interface PlanInput {
  name: string;
  price: number;
  durationDays: number;
  active?: boolean;
}

export interface Trainer {
  id: string;
  name: string;
  phone: string;
  specialization: string;
  notes: string;
}

export interface TrainerInput {
  name: string;
  phone?: string;
  specialization?: string;
  notes?: string;
}

export interface Membership {
  id: string;
  planId: string;
  startDate: string;
  endDate: string;
}

export interface Member {
  id: string;
  memberCode: string;
  name: string;
  phone: string;
  gender: string;
  dob?: string | null;
  address: string;
  emergencyContact: string;
  joinDate: string;
  trainerId?: string | null;
  notes: string;
  archived: boolean;
  membership?: Membership | null;
}

export interface MemberInput {
  name: string;
  phone?: string;
  gender?: string;
  dob?: string | null;
  address?: string;
  emergencyContact?: string;
  trainerId?: string | null;
  notes?: string;
}

export interface MemberPatch {
  name?: string;
  phone?: string;
  gender?: string;
  dob?: string | null;
  address?: string;
  emergencyContact?: string;
  trainerId?: string | null;
  notes?: string;
}

export interface MembershipInput {
  planId: string;
  startDate: string;
  endDate: string;
}

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  Cash: 'Cash',
  Easypaisa: 'Easypaisa',
  JazzCash: 'JazzCash',
  Bank_Transfer: 'Bank Transfer',
  Card: 'Card',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  completed: 'completed',
  voided: 'voided',
} as const;

export interface Payment {
  id: string;
  receiptNumber: string;
  transactionId: string;
  memberId: string;
  membershipId?: string | null;
  amount: number;
  method: PaymentMethod;
  date: string;
  notes: string;
  status: PaymentStatus;
}

export type PaymentInputMethod = typeof PaymentInputMethod[keyof typeof PaymentInputMethod];


export const PaymentInputMethod = {
  Cash: 'Cash',
  Easypaisa: 'Easypaisa',
  JazzCash: 'JazzCash',
  Bank_Transfer: 'Bank Transfer',
  Card: 'Card',
} as const;

export interface PaymentInput {
  memberId: string;
  membershipId?: string | null;
  amount: number;
  method: PaymentInputMethod;
  notes?: string;
}

export interface VoidPaymentInput {
  reason: string;
}

export type AttendanceSource = typeof AttendanceSource[keyof typeof AttendanceSource];


export const AttendanceSource = {
  manual: 'manual',
  biometric: 'biometric',
} as const;

export interface Attendance {
  id: string;
  memberId: string;
  checkIn: string;
  checkOut?: string | null;
  source: AttendanceSource;
}

export interface AttendanceInput {
  memberId: string;
}

/**
 * Not signed in
 */
export type UnauthorizedResponse = ApiErrorBody;

/**
 * Not found
 */
export type NotFoundResponse = ApiErrorBody;