# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: plans
    description: Membership plan operations
  - name: trainers
    description: Trainer operations
  - name: members
    description: Member operations
  - name: payments
    description: Payment operations
  - name: attendance
    description: Attendance operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List plans for the signed-in gym
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /plans/{id}/toggle:
    patch:
      operationId: togglePlan
      tags: [plans]
      summary: Toggle a plan's active flag
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /trainers:
    get:
      operationId: listTrainers
      tags: [trainers]
      summary: List trainers for the signed-in gym
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTrainer
      tags: [trainers]
      summary: Add a trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members for the signed-in gym
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createMember
      tags: [members]
      summary: Add a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get one member
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member's details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /members/{id}/archive:
    patch:
      operationId: archiveMember
      tags: [members]
      summary: Archive or restore a member
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /members/{id}/membership:
    patch:
      operationId: upsertMembership
      tags: [members]
      summary: Assign or renew a member's membership
      description: >-
        Upserts the member's single current membership (one active
        membership per member, no renewal history in this table -
        payments.membershipId is what records the renewal trail).
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments for the signed-in gym
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a full payment
      description: >-
        Full payment only - no partial payments. Server generates
        receiptNumber/transactionId and sets status to completed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /payments/{id}/void:
    patch:
      operationId: voidPayment
      tags: [payments]
      summary: Void a payment
      description: >-
        Void-only - payments are never edited or deleted, status only
        ever moves completed -> voided.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoidPaymentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records for the signed-in gym
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Record a manual check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/{id}/checkout:
    patch:
      operationId: checkoutAttendance
      tags: [attendance]
      summary: Record a check-out
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorBody"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorBody"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiErrorBody:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ArchiveInput:
      type: object
      properties:
        archived:
          type: boolean
      required: [archived]
    Plan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        durationDays:
          type: number
        active:
          type: boolean
      required: [id, name, price, durationDays, active]
    PlanInput:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationDays:
          type: number
        active:
          type: boolean
      required: [name, price, durationDays]
    Trainer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        specialization:
          type: string
        notes:
          type: string
      required: [id, name, phone, specialization, notes]
    TrainerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        specialization:
          type: string
        notes:
          type: string
      required: [name]
    Membership:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [id, planId, startDate, endDate]
    Member:
      type: object
      properties:
        id:
          type: string
        memberCode:
          type: string
        name:
          type: string
        phone:
          type: string
        gender:
          type: string
        dob:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
        emergencyContact:
          type: string
        joinDate:
          type: string
          format: date-time
        trainerId:
          type: string
          nullable: true
        notes:
          type: string
        archived:
          type: boolean
        membership:
          anyOf:
            - $ref: "#/components/schemas/Membership"
            - type: "null"
      required:
        [id, memberCode, name, phone, gender, address, emergencyContact, joinDate, notes, archived]
    MemberInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        gender:
          type: string
        dob:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
        emergencyContact:
          type: string
        trainerId:
          type: string
          nullable: true
        notes:
          type: string
      required: [name]
    MemberPatch:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        gender:
          type: string
        dob:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
        emergencyContact:
          type: string
        trainerId:
          type: string
          nullable: true
        notes:
          type: string

    MembershipInput:
      type: object
      properties:
        planId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [planId, startDate, endDate]
    Payment:
      type: object
      properties:
        id:
          type: string
        receiptNumber:
          type: string
        transactionId:
          type: string
        memberId:
          type: string
        membershipId:
          type: string
          nullable: true
        amount:
          type: number
        method:
          type: string
          enum: [Cash, Easypaisa, JazzCash, "Bank Transfer", Card]
        date:
          type: string
          format: date-time
        notes:
          type: string
        status:
          type: string
          enum: [completed, voided]
      required:
        [id, receiptNumber, transactionId, memberId, amount, method, date, notes, status]
    PaymentInput:
      type: object
      properties:
        memberId:
          type: string
        membershipId:
          type: string
          nullable: true
        amount:
          type: number
        method:
          type: string
          enum: [Cash, Easypaisa, JazzCash, "Bank Transfer", Card]
        notes:
          type: string
      required: [memberId, amount, method]
    VoidPaymentInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    Attendance:
      type: object
      properties:
        id:
          type: string
        memberId:
          type: string
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
          enum: [manual, biometric]
      required: [id, memberId, checkIn, source]
    AttendanceInput:
      type: object
      properties:
        memberId:
          type: string
      required: [memberId]