# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceSource } from './attendanceSource';

export interface Attendance {
  id: string;
  memberId: string;
  checkIn: Date;
  checkOut?: Date | null;
  source: AttendanceSource;
}