# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from './paymentMethod';
import type { PaymentStatus } from './paymentStatus';

export interface Payment {
  id: string;
  receiptNumber: string;
  transactionId: string;
  memberId: string;
  membershipId?: string | null;
  amount: number;
  method: PaymentMethod;
  date: Date;
  notes: string;
  status: PaymentStatus;
}