# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentInputMethod } from './paymentInputMethod';

export interface PaymentInput {
  memberId: string;
  membershipId?: string | null;
  amount: number;
  method: PaymentInputMethod;
  notes?: string;
}